// src/store/useCart.types.ts
export type ProductType = {
  id: number;
  title: string;
  price: number;
  description: string;
  category: string;
  image: string;
  rating: Rating;
  count?: number;
  total?: number;
};

export type Rating = {
  rate: number;
  count: number;
};

export type CartItem = {
  [key: string | number]: {
    count: number;
  };
};

export type UseCart = {
  cartItems: CartItem;
  products: ProductType[];
  loading: boolean;
  error: boolean; // Generated by https://quicktype.io

  fetchProducts: () => void;
  deleteFromCart: (id: number) => void;
  addToCart: (id: number) => void;
  removeFromCart: (id: number) => void;
};